Use dependency injection
