+++ CSharp/FizzBuzzKata/FizzBuzzKata.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.30907.101
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FizzBuzzKata", "FizzBuzzKata\FizzBuzzKata.csproj", "{B7ECC89F-94FB-457D-9277-8020C1B2D04F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FizzBuzzLogic", "FizzBuzzLogic\FizzBuzzLogic.csproj", "{6DBCE5E7-75AF-4909-9CD5-1302FE25355B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{B7ECC89F-94FB-457D-9277-8020C1B2D04F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B7ECC89F-94FB-457D-9277-8020C1B2D04F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B7ECC89F-94FB-457D-9277-8020C1B2D04F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6DBCE5E7-75AF-4909-9CD5-1302FE25355B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6DBCE5E7-75AF-4909-9CD5-1302FE25355B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6DBCE5E7-75AF-4909-9CD5-1302FE25355B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6DBCE5E7-75AF-4909-9CD5-1302FE25355B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
