+++ CSharp/FizzBuzzKata/FizzBuzzKata/FizzBuzzKata.csproj
@@ -9,7 +9,11 @@
   <ItemGroup>
     <PackageReference Include="NUnit" Version="3.12.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="3.16.1" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0"/>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\FizzBuzzLogic\FizzBuzzLogic.csproj" />
   </ItemGroup>
 
 </Project>
