+++ CSharp/FizzBuzzKata/FizzBuzzLogic/Contract/IBuzzConverter.cs
@@ -1,0 +1,7 @@
+namespace FizzBuzzLogic.Contract
+{
+    public interface IBuzzConverter
+    {
+        string Convert(int parameter);
+    }
+}
