+++ CSharp/FizzBuzzKata/FizzBuzzLogic/Contract/IConverter.cs
@@ -1,0 +1,7 @@
+namespace FizzBuzzLogic.Contract
+{
+    public interface IConverter
+    {
+        string Convert(int parameter);
+    }
+}
