+++ CSharp/FizzBuzzKata/FizzBuzzLogic/Contract/IFizzConverter.cs
@@ -1,0 +1,7 @@
+namespace FizzBuzzLogic.Contract
+{
+    public interface IFizzConverter
+    {
+        string Convert(int parameter);
+    }
+}
