+++ CSharp/FizzBuzzKata/FizzBuzzLogic/Implementation/BuzzConverter.cs
@@ -1,0 +1,12 @@
+using FizzBuzzLogic.Contract;
+
+namespace FizzBuzzLogic.Implementation
+{
+    public class BuzzConverter : IBuzzConverter
+    {
+        public string Convert(int parameter)
+        {
+            return parameter % 5 == 0 ? "BUZZ" : string.Empty;
+        }
+    }
+}
